New SPA. Add GooglePlayUnavailableException.cs
