+++ Pub/PubVersion.json
@@ -1,6 +1,6 @@
 {
-  "Version": "5.0.633",
+  "Version": "5.0.634",
   "BumpTime": "2025-01-02T09:10:23.0535339Z",
-  "Prerelease": true,
+  "Prerelease": false,
   "DeprecatedVersion": "4.0.00"
 }
