+++ Src/AppLib/VpnHood.AppLib.Android.GooglePlay/Exceptions/GooglePlayUnavailableException.cs
@@ -1,0 +1,3 @@
+﻿namespace VpnHood.AppLib.Droid.GooglePlay;
+
+public class GooglePlayUnavailableException() : Exception("GooglePlay is unavailable");
